% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{AddInfo}
\alias{AddInfo}
\alias{AddInfo.IST}
\title{Add information to IST object}
\usage{
AddInfo(IST_obj, info_key, sub_key, info_value, ...)

\method{AddInfo}{IST}(
  IST_obj = NULL,
  info_key = "comment_info",
  sub_key = NULL,
  info_value = NULL,
  ...
)
}
\arguments{
\item{IST_obj}{An IST object}

\item{info_key}{Character, information category (default: "comment_info")}

\item{sub_key}{Character, specific sub-key to add to}

\item{info_value}{Value to append}

\item{...}{Additional arguments passed to methods}
}
\value{
Modified IST object
}
\description{
Appends information to existing slots in the IST_info, primarily for
adding comments and notes.
}
\examples{
\dontrun{
# Add a comment about analysis parameters
IST_obj <- AddInfo(ist_obj,
                   info_key = "comment_info",
                   sub_key = "parameters",
                   info_value = "Used DBSCAN with eps=5, minPts=10")
}

}
