% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{AddSSNicheCells}
\alias{AddSSNicheCells}
\alias{AddSSNicheCells.IST}
\title{Add columns to SingleSampNiche cells}
\usage{
AddSSNicheCells(IST_obj, loop_id, meta_key, select_colnm, niche_key, ...)

\method{AddSSNicheCells}{IST}(
  IST_obj = NULL,
  loop_id = "LoopAllSamp",
  meta_key = NULL,
  select_colnm = NULL,
  niche_key = NULL,
  ...
)
}
\arguments{
\item{IST_obj}{An IST object}

\item{loop_id}{Character vector, sample identifiers}

\item{meta_key}{Character, metadata key containing new data}

\item{select_colnm}{Character vector, columns to add}

\item{niche_key}{Character, niche key to modify}

\item{...}{Additional arguments passed to methods}
}
\value{
Modified IST object with updated niche cell data
}
\description{
Adds new columns to the cell data in SingleSampNiche objects.
}
\examples{
\dontrun{
# Add new annotation columns to niche cells
IST_obj <- AddSSNicheCells(ist_obj,
                           loop_id = "LoopAllSamp",
                           meta_key = "custom_annotation",
                           select_colnm = c("cell_type", "confidence"),
                           niche_key = "niche_virulence")
}

}
