% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{AddSSNicheGenes}
\alias{AddSSNicheGenes}
\alias{AddSSNicheGenes.IST}
\title{Add annotations to SingleSampNiche genes}
\usage{
AddSSNicheGenes(IST_obj, ...)

\method{AddSSNicheGenes}{IST}(
  IST_obj = NULL,
  gene = NULL,
  label = NULL,
  add_colnm = NULL,
  loop_id = "LoopAllSamp",
  niche_key = NULL,
  ...
)
}
\arguments{
\item{IST_obj}{An IST object}

\item{...}{Additional arguments passed to methods}

\item{gene}{Character vector, gene names to annotate}

\item{label}{Character vector, labels to assign to genes (optional)}

\item{add_colnm}{Character, name of new column to add}

\item{loop_id}{Character vector, sample identifiers}

\item{niche_key}{Character, niche key to modify}
}
\value{
Modified IST object with updated niche gene data
}
\description{
Adds new columns to the gene data in SingleSampNiche objects.
}
\examples{
\dontrun{
# Add gene family annotations
IST_obj <- AddSSNicheGenes(ist_obj,
                           gene = c("gene1", "gene2", "gene3"),
                           label = c("toxin", "toxin", "adhesin"),
                           add_colnm = "gene_family",
                           niche_key = "niche_virulence")
}

}
