% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{as.Seurat.IST}
\alias{as.Seurat.IST}
\title{Convert IST object to Seurat object}
\usage{
as.Seurat.IST(IST_obj)
}
\arguments{
\item{IST_obj}{An IST object to be converted}
}
\value{
A Seurat object containing the same assay data and basic metadata
}
\description{
Converts an IST object back to a standard Seurat object, preserving all
assay data and basic metadata but removing IST-specific analysis structures.
}
\examples{
\dontrun{
# Convert IST object back to Seurat
seurat_obj <- as.Seurat.IST(ist_obj)
}

}
