% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSamp_Niche.R
\name{CalNichePredict}
\alias{CalNichePredict}
\title{Predict Niche Types or Properties}
\usage{
CalNichePredict(x)
}
\arguments{
\item{x}{Placeholder parameter for future expansion}
}
\value{
NULL (invisible)
}
\description{
Placeholder function for predictive modeling of niche properties based on
spatial and molecular features. Currently under development.
}
\details{
This function is designed to build predictive models for niche classification
or property estimation using machine learning approaches.
}
\examples{
\dontrun{
# Under development
CalNichePredict()
}
}
