% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetCoordInfo}
\alias{GetCoordInfo}
\alias{GetCoordInfo.IST}
\title{Get coordinate-related column names}
\usage{
GetCoordInfo(IST_obj)

\method{GetCoordInfo}{IST}(IST_obj = NULL)
}
\arguments{
\item{IST_obj}{An IST object}
}
\value{
Named vector of coordinate column names
}
\description{
Convenience function to retrieve all coordinate-related column names
from an IST object.
}
\examples{
\dontrun{
coord_info <- GetCoordInfo(ist_obj)
}

}
