% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetInfo}
\alias{GetInfo}
\alias{GetInfo.IST}
\title{Get information from IST object}
\usage{
GetInfo(IST_obj, info_key, ...)

\method{GetInfo}{IST}(IST_obj = NULL, info_key = NULL, sub_key = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{info_key}{Character, main information category (e.g., "samp_info", "data_info")}

\item{...}{Additional arguments passed to methods}

\item{sub_key}{Character vector, specific sub-keys to retrieve (optional)}
}
\value{
Requested information from the IST object
}
\description{
Retrieves specific information from the IST object's IST_info slot.
}
\examples{
\dontrun{
# Get sample column name
samp_colnm <- GetInfo(ist_obj, info_key = "data_info", sub_key = "samp_colnm")

# Get all data information
data_info <- GetInfo(ist_obj, info_key = "data_info")
}

}
