% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetMetaInfo}
\alias{GetMetaInfo}
\alias{GetMetaInfo.IST}
\title{Get metadata information}
\usage{
GetMetaInfo(IST_obj, ...)

\method{GetMetaInfo}{IST}(IST_obj = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{...}{Additional arguments passed to methods}
}
\value{
Data frame containing metadata tracking information
}
\description{
Retrieves the metadata tracking information data frame from an IST object.
}
\examples{
\dontrun{
meta_info <- GetMetaInfo(ist_obj)
}

}
