% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetMSNicheCells}
\alias{GetMSNicheCells}
\alias{GetMSNicheCells.IST}
\title{Get MultiSampNiche cells}
\usage{
GetMSNicheCells(IST_obj, loop_id, niche_key, ...)

\method{GetMSNicheCells}{IST}(
  IST_obj = NULL,
  loop_id = "LoopAllMulti",
  niche_key = NULL,
  ...
)
}
\arguments{
\item{IST_obj}{An IST object}

\item{loop_id}{Character vector, multi-sample analysis identifiers}

\item{niche_key}{Character, niche key to retrieve}

\item{...}{Additional arguments passed to methods}
}
\value{
List of cell data frames, one per multi-sample analysis
}
\description{
Retrieves cell-level data from MultiSampNiche objects for specified multi-sample analyses.
}
\examples{
\dontrun{
multi_cells <- GetMSNicheCells(ist_obj,
                               loop_id = c("comparison1", "comparison2"),
                               niche_key = "niche_virulence")
}

}
