% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetMSNicheInfo}
\alias{GetMSNicheInfo}
\alias{GetMSNicheInfo.IST}
\title{Get MultiSampNiche information}
\usage{
GetMSNicheInfo(IST_obj, loop_id, niche_key, ...)

\method{GetMSNicheInfo}{IST}(IST_obj = NULL, loop_id = "LoopAllMulti", niche_key = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{loop_id}{Character vector, multi-sample analysis identifiers}

\item{niche_key}{Character, niche key to retrieve}

\item{...}{Additional arguments passed to methods}
}
\value{
List of niche information data frames, one per multi-sample analysis
}
\description{
Retrieves niche information from MultiSampNiche objects for specified multi-sample analyses.
}
\examples{
\dontrun{
multi_info <- GetMSNicheInfo(ist_obj,
                             loop_id = c("comparison1", "comparison2"),
                             niche_key = "niche_virulence")
}

}
