% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetSSNicheGenes}
\alias{GetSSNicheGenes}
\alias{GetSSNicheGenes.IST}
\title{Get SingleSampNiche genes}
\usage{
GetSSNicheGenes(IST_obj, loop_id, niche_key, ...)

\method{GetSSNicheGenes}{IST}(IST_obj = NULL, loop_id = "LoopAllSamp", niche_key = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{loop_id}{Character vector, sample identifiers}

\item{niche_key}{Character, niche key to retrieve}

\item{...}{Additional arguments passed to methods}
}
\value{
List of gene data frames, one per sample
}
\description{
Retrieves gene-level data from SingleSampNiche objects for specified samples.
}
\examples{
\dontrun{
niche_genes <- GetSSNicheGenes(ist_obj,
                               loop_id = c("sample1", "sample2"),
                               niche_key = "niche_virulence")
}

}
