% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{GetSSNicheInfo}
\alias{GetSSNicheInfo}
\alias{GetSSNicheInfo.IST}
\title{Get SingleSampNiche information}
\usage{
GetSSNicheInfo(IST_obj, loop_id, niche_key, ...)

\method{GetSSNicheInfo}{IST}(IST_obj = NULL, loop_id = "LoopAllSamp", niche_key = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{loop_id}{Character vector, sample identifiers}

\item{niche_key}{Character, niche key to retrieve}

\item{...}{Additional arguments passed to methods}
}
\value{
List of niche information data frames, one per sample
}
\description{
Retrieves niche information from SingleSampNiche objects for specified samples.
}
\examples{
\dontrun{
niche_info <- GetSSNicheInfo(ist_obj,
                             loop_id = c("sample1", "sample2"),
                             niche_key = "niche_virulence")
}

}
