% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\docType{class}
\name{IST-class}
\alias{IST-class}
\alias{IST}
\title{The IST Class}
\description{
The IST (Integrated Single-cell Transcriptomics) object extends the Seurat
class to add specialized functionality for analyzing infection and spatial
transcriptomics data. It inherits all functionality from Seurat objects while
adding additional slots for tracking infection-related analyses.
}
\section{Slots}{

\describe{
\item{\code{IST_analysis}}{An \code{\link{IST_analysis}} object containing all
infection-specific analysis results and metadata}

\item{\code{assays}}{A list of assays for this project (inherited from Seurat)}

\item{\code{meta.data}}{Contains meta-information about each cell (inherited from Seurat)}

\item{\code{active.assay}}{Name of the active, or default, assay (inherited from Seurat)}

\item{\code{active.ident}}{The active cluster identity (inherited from Seurat)}

\item{\code{graphs}}{A list of \code{\link[SeuratObject]{Graph}} objects (inherited from Seurat)}

\item{\code{neighbors}}{Neighbor graphs (inherited from Seurat)}

\item{\code{reductions}}{A list of dimensional reduction objects (inherited from Seurat)}

\item{\code{images}}{A list of spatial image objects (inherited from Seurat)}

\item{\code{project.name}}{Name of the project (inherited from Seurat)}

\item{\code{misc}}{A list of miscellaneous information (inherited from Seurat)}

\item{\code{version}}{Version of Seurat this object was built under (inherited from Seurat)}

\item{\code{commands}}{A list of logged commands (inherited from Seurat)}

\item{\code{tools}}{A list of miscellaneous data generated by other tools (inherited from Seurat)}
}}

\examples{
\dontrun{
# Create an IST object from a Seurat object
ist_obj <- as.IST(
  seurat_obj = seurat_object,
  samp_colnm = "sample_id",
  celltype_colnm = "cell_type",
  host_org = "human",
  pathogen_grp = "virus",
  pathogen_org = "SARS-CoV-2",
  data_format = "StereoSeq"
)
}

}
\seealso{
Other IST-classes: 
\code{\link{IST_analysis-class}}
}
\concept{IST-classes}
