% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{IST_info}
\alias{IST_info}
\title{Get IST package information}
\usage{
IST_info()
}
\value{
A list containing package information with the following components:
\itemize{
\item \code{version}: Package version
\item \code{functions}: List of available functions by category
\item \code{supported_formats}: Supported spatial transcriptomics formats
\item \code{supported_hosts}: Supported host species
\item \code{supported_pathogens}: Supported pathogen types
\item \code{score_methods}: Available scoring methods
\item \code{blur_methods}: Available blurring methods
\item \code{options}: Current package options
}
}
\description{
Returns information about the IST package including version, available functions,
and current package options.
}
\examples{
\dontrun{
info <- IST_info()
print(info$version)
print(info$supported_formats)
}
}
