% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{RemoveMetaColumn}
\alias{RemoveMetaColumn}
\alias{RemoveMetaColumn.IST}
\title{Remove columns from existing metadata}
\usage{
RemoveMetaColumn(IST_obj, ...)

\method{RemoveMetaColumn}{IST}(IST_obj = NULL, meta_key = NULL, remove_colnm = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{...}{Additional arguments passed to methods}

\item{meta_key}{Character, metadata key to modify}

\item{remove_colnm}{Character vector, column names to remove}
}
\value{
Modified IST object with removed metadata columns
}
\description{
Removes specified columns from an existing metadata data frame in the IST object.
}
\examples{
\dontrun{
# Remove temporary columns
IST_obj <- RemoveMetaColumn(ist_obj,
                            meta_key = "raw",
                            remove_colnm = c("temp_score", "temp_label"))
}

}
