% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{RemoveMetaData}
\alias{RemoveMetaData}
\alias{RemoveMetaData.IST}
\title{Remove metadata from IST object}
\usage{
RemoveMetaData(IST_obj, meta_key, ...)

\method{RemoveMetaData}{IST}(IST_obj = NULL, meta_key = NULL, ...)
}
\arguments{
\item{IST_obj}{An IST object}

\item{meta_key}{Character vector, metadata keys to remove}

\item{...}{Additional arguments passed to methods}
}
\value{
Modified IST object with removed metadata
}
\description{
Removes specified metadata from the IST object's meta_data_record.
}
\examples{
\dontrun{
# Remove temporary metadata
IST_obj <- RemoveMetaData(ist_obj, meta_key = "temp_analysis")
}

}
