% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IST_Class.R
\name{SetInfo}
\alias{SetInfo}
\alias{SetInfo.IST}
\title{Set information in IST object}
\usage{
SetInfo(IST_obj, info_key, sub_key, info_value, ...)

\method{SetInfo}{IST}(
  IST_obj = NULL,
  info_key = NULL,
  sub_key = NULL,
  info_value = NULL,
  ...
)
}
\arguments{
\item{IST_obj}{An IST object}

\item{info_key}{Character, main information category}

\item{sub_key}{Character, specific sub-key to set (optional)}

\item{info_value}{Value to set}

\item{...}{Additional arguments passed to methods}
}
\value{
Modified IST object
}
\description{
Modifies specific information in the IST object's IST_info slot.
}
\examples{
\dontrun{
# Update project description
IST_obj <- SetInfo(ist_obj,
                   info_key = "project_info",
                   sub_key = "description",
                   info_value = "Updated project description")
}

}
